+++ on_audio_query/lib/details/on_audio_query_controller.dart
@@ -41,12 +41,14 @@ class OnAudioQuery {
   /// Will check and request, if necessary, all required permissions.
   ///
   /// **OBS: Will always return true on web platform.**
-  Future<bool> checkAndRequest() async {
+  Future<bool> checkAndRequest({bool retryRequest = false}) async {
     if (kIsWeb) return true;
 
     bool hasPermission = await platform.permissionsStatus();
     if (!hasPermission) {
-      hasPermission = await platform.permissionsRequest();
+      hasPermission = await platform.permissionsRequest(
+        retryRequest: retryRequest,
+      );
     }
 
     return hasPermission;
@@ -567,7 +569,8 @@ class OnAudioQuery {
   /// Important:
   ///
   /// * This method will always return a bool.
-  /// * If return true `[READ]` and `[WRITE]` permissions is Granted, else `[READ]` and `[WRITE]` is Denied.
+  /// * If return true `[READ]` and `[WRITE]` permissions is Granted,
+  /// else `[READ]` and `[WRITE]` is Denied.
   ///
   /// Platforms:
   ///
@@ -585,7 +588,8 @@ class OnAudioQuery {
   /// Important:
   ///
   /// * This method will always return a bool.
-  /// * If return true `[READ]` and `[WRITE]` permissions is Granted, else `[READ]` and `[WRITE]` is Denied.
+  /// * If return true `[READ]` and `[WRITE]` permissions is Granted,
+  /// else `[READ]` and `[WRITE]` is Denied.
   ///
   /// Platforms:
   ///
@@ -594,8 +598,8 @@ class OnAudioQuery {
   /// | `✔️` | `✔️` | `❌` | <br>
   ///
   /// See more about [platforms support](https://github.com/LucJosin/on_audio_query/blob/main/PLATFORMS.md)
-  Future<bool> permissionsRequest() async {
-    return platform.permissionsRequest();
+  Future<bool> permissionsRequest({bool retryRequest = false}) async {
+    return platform.permissionsRequest(retryRequest: retryRequest);
   }
 
   // Device Information