+++ on_audio_query_platform_interface/lib/method_channel_on_audio_query.dart
@@ -321,9 +321,12 @@ class MethodChannelOnAudioQuery extends OnAudioQueryPlatform {
   }
 
   @override
-  Future<bool> permissionsRequest() async {
+  Future<bool> permissionsRequest({bool retryRequest = false}) async {
     final bool resultRequest = await _channel.invokeMethod(
       "permissionsRequest",
+      {
+        "retryRequest": retryRequest,
+      },
     );
     return resultRequest;
   }